+++ extensions/dear-imnodes-sys/build.rs
@@ -106,6 +106,8 @@ fn generate_bindings(
         .allowlist_function("getIOKeyCtrlPtr")
         .allowlist_function("imnodes_getIOKeyShiftPtr")
         .allowlist_function("imnodes_getIOKeyAltPtr")
+        .allowlist_function("imnodes_EditorContextResetToDefault")
+        .allowlist_function("imnodes_EditorContextResetToDefaultIfCurrent")
         .allowlist_type("ImNodes.*")
         .allowlist_type("ImWchar32")
         .allowlist_var("ImNodes.*")