+++ extensions/dear-imnodes-sys/shim/imnodes_extra.cpp
@@ -1,6 +1,17 @@
-#include "imgui.h"
+#include "imnodes_internal.h"
 extern "C" {
 bool* imnodes_getIOKeyShiftPtr() { return &ImGui::GetIO().KeyShift; }
 bool* imnodes_getIOKeyAltPtr() { return &ImGui::GetIO().KeyAlt; }
+void imnodes_EditorContextResetToDefault()
+{
+    if (GImNodes != nullptr) {
+        GImNodes->EditorCtx = GImNodes->DefaultEditorCtx;
+    }
 }
-
+void imnodes_EditorContextResetToDefaultIfCurrent(ImNodesEditorContext* editor)
+{
+    if (GImNodes != nullptr && GImNodes->EditorCtx == editor) {
+        GImNodes->EditorCtx = GImNodes->DefaultEditorCtx;
+    }
+}
+}